Moved all db code to db.ts
